// src/app/@core/dtos/post-agent-details.dto.ts
export interface PostAgentDetailsDto {
    bvn: string;
    dateOfBirth: string;
    address: string;
    businessName: string;
    officeAddress: string;
    officeState: string;
    officeLga: string;
    gender: string;
    nokFullName: string;
    nokEmail: string;
    nokPhone: string;
    guarantorsFullName: string;
    guarantorsProfession: string;
    guarantorsEmail: string;
    guarantorsResidence: string;
    guarantorsOffice: string;
    guarantorsPhone: string;
    guarantorsRelationship: string;
}